removed the placeholder from auth form fields
